refactor: update imports and ESLint configuration
- Refactor imports to utilize the new ifc-utils module across various components and utilities.
- Update ESLint configuration to disable specific rules
