+++ .eslintrc.json
@@ -1,6 +1,9 @@
 {
   "extends": [
-    "next/core-web-vitals",
-    "next/typescript"
-  ]
+    "next/core-web-vitals"
+  ],
+  "rules": {
+    "react/no-unescaped-entities": "off",
+    "react-hooks/exhaustive-deps": "off"
+  }
 }
