+++ lib/ifc/index.ts
@@ -1,10 +1,12 @@
 // Export all IFC utilities from a single file
-// Using absolute imports to avoid path resolution issues
-export * from "@/lib/ifc/ifc-loader"
 export * from "@/lib/ifc/geometry-utils"
 export * from "@/lib/ifc/filter-utils"
 export * from "@/lib/ifc/quantity-utils"
 export * from "@/lib/ifc/property-utils"
 export * from "@/lib/ifc/spatial-utils"
 export * from "@/lib/ifc/analysis-utils"
+export * from "@/lib/ifc/cluster-utils"
+export * from "@/lib/ifc/file-uploader"
+export * from "@/lib/ifc/viewer-manager"
+export * from "@/lib/ifc/viewer-utils"
